requirements.txt
numpy
pandas
websockets
paho-mqtt
opcua
pymodbus
pyserial
psutil
